requirements.txt
# Core dependencies
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
mistapi>=0.57.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Configuration and logging
pyyaml>=6.0
colorlog>=6.7.0

# Web framework for dashboard
flask>=2.3.0
flask-cors>=4.0.0

# Database support
sqlalchemy>=2.0.0

# Scheduling and task management
apscheduler>=3.10.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Notification systems
slack-sdk>=3.21.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
requests-mock>=1.11.0

# Data visualization (optional)
plotly>=5.15.0
dash>=2.11.0

# Environment and configuration
python-decouple>=3.8

# JSON handling
jsonschema>=4.18.0

# HTTP server for webhooks
waitress>=2.1.0

# Retry mechanisms
tenacity>=8.2.0

# Time series data (optional)
influxdb-client>=1.37.0

# Mist Wireless Troubleshooting dependencies
ipaddress>=1.0.23
